requirements.txt
# Core
Django>=4.2,<5.0
python-dotenv>=1.0.0

# Server (production WSGI server)
gunicorn>=21.2.0

# Static files (serves CSS/JS in production without a separate web server)
whitenoise>=6.6.0

# Database (PostgreSQL for production)
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# AI / API
openai>=1.0.0
requests>=2.31.0

# Document generation
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.0
reportlab>=4.0.0

# Image handling
Pillow>=10.0.0

# REST API
djangorestframework>=3.15.0
django-filter>=23.0

# Admin UI
django-jazzmin>=2.6.0

# Cloud file storage
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
